Fix Rails 7.2 skip_before_action compatibility
Rails 7.2 changed skip_before_action to raise an error by default
when the callback doesn't exist. The HealthController was trying to
skip :authenticate_user! which is not defined as a before_action in
ApplicationController (Devise defines it in specific controllers).

Add raise: false to allow skipping callbacks that may not exist.